Add Privacy policy. Fixes #20
The online version of this can be viewed at https://github.com/justdave/MCStatus/blob/master/PRIVACY.md

Both PRIVACY.md and res/raw/privacy.txt will need to be separately updated if changes need to be made. The content of the two files are identical, but they need different file extensions for Android and GitHub to display them properly.